+++ packages/libraries/csv/src/streamToArrayBuffer.ts
@@ -1,0 +1,18 @@
+// https://stackoverflow.com/questions/40385133/retrieve-data-from-a-readablestream-object
+function concatArrayBuffers(chunks: Uint8Array[]): Uint8Array {
+  const result = new Uint8Array(chunks.reduce((a, c) => a + c.length, 0));
+  let offset = 0;
+  for (const chunk of chunks) {
+    result.set(chunk, offset);
+    offset += chunk.length;
+  }
+  return result;
+}
+
+export async function streamToArrayBuffer(stream: ReadableStream<Uint8Array>): Promise<Uint8Array> {
+  const chunks: Uint8Array[] = [];
+  for await (const chunk of stream) {
+    chunks.push(chunk);
+  }
+  return concatArrayBuffers(chunks);
+}
